+++ servers/physics_2d/physics_server_2d.cpp
@@ -495,7 +495,7 @@ TypedArray<uint64_t> PhysicsTestMotionParameters2D::get_exclude_objects() const 
 
 	int object_index = 0;
 	for (ObjectID object_id : parameters.exclude_objects) {
-		exclude[object_index++] = object_id;
+		exclude[object_index++] = uint64_t(object_id);
 	}
 
 	return exclude;
